app/backend/requirements.txt
functions-framework==3.*
google-cloud-aiplatform[agent_engines]>=1.40
